+++ libs/common/src/lib/personal-finance-tools.ts
@@ -1591,6 +1591,26 @@ export const personalFinanceTools: Product[] = [
     url: 'https://www.navexa.com'
   },
   {
+    categories: [
+      'BUDGETING',
+      'CRYPTOCURRENCY',
+      'ETF_TRACKING',
+      'FINANCIAL_PLANNING',
+      'NET_WORTH_TRACKING',
+      'STOCK_TRACKING'
+    ],
+    hasFreePlan: true,
+    hasSelfHostingAbility: false,
+    key: 'neontra',
+    languages: ['English'],
+    name: 'Neontra',
+    origin: 'CA',
+    platforms: ['ANDROID', 'IOS', 'WEB'],
+    pricingPerYear: '$75',
+    slogan: 'Everything you own. Everything you owe. Everything in one place.',
+    url: 'https://neontra.com'
+  },
+  {
     categories: ['INVESTMENT_RESEARCH', 'STOCK_TRACKING'],
     hasFreePlan: true,
     hasSelfHostingAbility: false,